backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==10.2.0
torch==2.1.2
diffusers==0.25.0
transformers==4.36.2
accelerate==0.25.0
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
hypothesis==6.92.1
